Set the license to Apache 2
When I have first changed the the License, for an unknown reason I have
specified inside the commit message that I have setup an MIT license
while I was pushing the Apache 2 version. This is a human error on my
side.

This commit have for only goal to replace the old commit message in
github in order to avoid confusion.